app-service/requirements.txt
# ── Core Framework ────────────────────────────────────────────
Flask==3.1.3                # web framework
Flask-SQLAlchemy==3.1.1     # database ORM
Flask-Login==0.6.3          # user sessions
Flask-WTF==1.2.2            # forms + CSRF
Flask-Bcrypt==1.0.1         # password hashing
WTForms==3.2.1              # form validators

# ── Database ──────────────────────────────────────────────────
psycopg2-binary==2.9.11     # PostgreSQL driver

# ── HTTP Client ───────────────────────────────────────────────
requests==2.32.3            # HTTP calls to catalogue-service

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.2.1        # reads .env file
email-validator==2.3.0      # validates email fields
gunicorn==25.1.0            # production server

# ── Testing ───────────────────────────────────────────────────
pytest==8.3.4               # test runner
pytest-flask==1.3.0         # Flask test helpers
